+++ src/services/index.ts
@@ -1,2 +1,3 @@
 export * from './configService';
 export * from './tasksService';
+export * from './resouceService';